some upgrades

+++ src/redux/pasteSlice.jsx
@@ -52,17 +52,19 @@ export const pasteSlice = createSlice({
       localStorage.removeItem("pastes");
     },
     removeFromPastes: (state, action) => {
-      const pasteId = action.payload;
-
-      const index = state.pastes.findIndex((item) => item._id === pasteId);
-
-      if (index >= 0) {
-        state.pastes.splice(index, 1);
-
-        localStorage.setItem("pastes", JSON.stringify(state.pastes));
+      if(confirm("Delete This Paste..?")){
+        const pasteId = action.payload;
+  
+        const index = state.pastes.findIndex((item) => item._id === pasteId);
+  
+        if (index >= 0) {
+          state.pastes.splice(index, 1);
+  
+          localStorage.setItem("pastes", JSON.stringify(state.pastes));
+        }
+  
+        toast.success("Paste Deleted");
       }
-
-      toast.success("Paste Deleted");
     },
   },
 });
